+++ src/lox/run.rs
@@ -2,7 +2,7 @@ use std::{collections::BTreeMap, fs, io};
 
 use crate::{
     cli::alerts::Alert,
-    errors::{Error, LoxError, SystemError},
+    errors::{Error, SystemError},
     lox::{
         interpreter::Interpreter,
         scanner::Scanner,
@@ -123,8 +123,7 @@ fn run(tokens: Vec<Token>) -> Result<(), Error> {
 
     let result = match Interpreter::evaluate(expr) {
         Ok(lit) => lit,
-        Err(_) => Error::from(LoxError::CustomError(0, "Runtime exception".to_string()))
-            .report_and_exit(1),
+        Err(runtime_err) => Error::from(runtime_err).report_and_exit(70),
     };
 
     println!("{:?}", result);